aice_api/requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
Werkzeug==2.3.6
langchain==0.0.240
openai==0.27.8
python-dotenv==1.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.6
Flask-JWT-Extended==4.5.2